Fixed battle stuff see description for more detail
The enemy stats should now properly transfer between the overworld and battle scene and it "Should" have stopped throwing the null reference exception when you fight an enemy.

I completely threw out the old GameDataHolder instance method for the enemy stats cuz it just didn't work properly between scenes and am now using a static class to hold the data. The GameManager is still there to keep track of which enemies are dead or not.

I have also rewritten the BattleSystem setup function it so that it will read the name of the enemy from the enemy stats data and load the corresponding correct sprite and optionally animation data in the battle scene as long as an entry for the sprites and animation controller are assigned within the Enemy Sprite Database object in the scripts folder.

I have replaced the broken prefabs with updated versions and gotten rid of a bunch of old redundant prefabs. Also the battle system now uses a generic enemy prefab with zeroed out stats that get overwritten with new stats when an enemy is encountered.

also adjusted a couple sprites positioning and made it so the player in the battle scene is the same sprite as the overworld.

there might be some more stuff i fixed that i cant remember right now.
tell me if anything breaks again cuz i'm pretty sure i should have fixed the battle related stuff and if its not fixed i'm convinced that this game is somehow haunted or something.

theres also a bunch of unnecessary debug logs that print to the console for random stuff that i added while i was trying to diagnose errors. i will get rid of them later.